+++ src/gui/app.rs
@@ -1,4 +1,5 @@
 use crate::{
+    cache::Cache,
     config::{Config, CustomGame, RootsConfig, Theme},
     gui::{
         backup_screen::BackupScreenComponent,
@@ -70,6 +71,7 @@ struct Progress {
 pub struct App {
     config: Config,
     manifest: Manifest,
+    cache: Cache,
     translator: Translator,
     operation: Option<OngoingOperation>,
     screen: Screen,
@@ -153,8 +155,8 @@ impl App {
                         &duplicates,
                     );
                 }
-                self.config.backup.recent_games.retain(|x| !games.contains(x));
-                self.config.save();
+                self.cache.backup.recent_games.retain(|x| !games.contains(x));
+                self.cache.save();
             }
             return Command::none();
         }
@@ -289,8 +291,8 @@ impl App {
                         &duplicates,
                     );
                 }
-                self.config.restore.recent_games.retain(|x| !games.contains(x));
-                self.config.save();
+                self.cache.restore.recent_games.retain(|x| !games.contains(x));
+                self.cache.save();
             }
             return Command::none();
         }
@@ -350,14 +352,11 @@ impl App {
         let mut failed = false;
 
         if full {
-            self.config.backup.recent_games.clear();
+            self.cache.backup.recent_games.clear();
         }
 
         for entry in &self.backup_screen.log.entries {
-            self.config
-                .backup
-                .recent_games
-                .insert(entry.scan_info.game_name.clone());
+            self.cache.backup.recent_games.insert(entry.scan_info.game_name.clone());
             if let Some(backup_info) = &entry.backup_info {
                 if !backup_info.successful() {
                     failed = true;
@@ -369,7 +368,7 @@ impl App {
             self.backup_screen.previewed_games.clear();
         }
 
-        self.config.save();
+        self.cache.save();
 
         if failed {
             self.modal_theme = Some(ModalTheme::Error {
@@ -386,11 +385,11 @@ impl App {
         let mut failed = false;
 
         if full {
-            self.config.restore.recent_games.clear();
+            self.cache.restore.recent_games.clear();
         }
 
         for entry in &self.restore_screen.log.entries {
-            self.config
+            self.cache
                 .restore
                 .recent_games
                 .insert(entry.scan_info.game_name.clone());
@@ -401,7 +400,7 @@ impl App {
             }
         }
 
-        self.config.save();
+        self.cache.save();
 
         if failed {
             self.modal_theme = Some(ModalTheme::Error {
@@ -463,7 +462,7 @@ impl Application for App {
     fn new(_flags: ()) -> (Self, Command<Message>) {
         let translator = Translator::default();
         let mut modal_theme: Option<ModalTheme> = None;
-        let config = match Config::load() {
+        let mut config = match Config::load() {
             Ok(x) => x,
             Err(x) => {
                 modal_theme = Some(ModalTheme::Error { variant: x });
@@ -472,6 +471,7 @@ impl Application for App {
             }
         };
         translator.set_language(config.language);
+        let cache = Cache::load().migrated(&mut config);
         let manifest = match Manifest::load_local() {
             Ok(y) => y,
             Err(_) => {
@@ -484,13 +484,14 @@ impl Application for App {
 
         (
             Self {
-                backup_screen: BackupScreenComponent::new(&config),
-                restore_screen: RestoreScreenComponent::new(&config),
+                backup_screen: BackupScreenComponent::new(&config, &cache),
+                restore_screen: RestoreScreenComponent::new(&config, &cache),
                 custom_games_screen: CustomGamesScreenComponent::new(&config),
                 other_screen: OtherScreenComponent::new(&config),
                 translator,
                 config,
                 manifest,
+                cache,
                 modal_theme,
                 ..Self::default()
             },
@@ -602,7 +603,7 @@ impl Application for App {
                             &self.backup_screen.duplicate_detector,
                             &duplicates,
                         );
-                        self.config.backup.recent_games.remove(&scan_info.game_name);
+                        self.cache.backup.recent_games.remove(&scan_info.game_name);
                     }
                 } else {
                     log::trace!(
@@ -657,7 +658,7 @@ impl Application for App {
                             &self.restore_screen.duplicate_detector,
                             &duplicates,
                         );
-                        self.config.restore.recent_games.remove(&scan_info.game_name);
+                        self.cache.restore.recent_games.remove(&scan_info.game_name);
                     }
                 } else {
                     log::trace!(