+++ src/prelude.rs
@@ -16,7 +16,6 @@ pub const CASE_INSENSITIVE_OS: bool = WINDOWS || MAC;
 const SKIP: &str = "<skip>";
 const APP_DIR_NAME: &str = "ludusavi";
 const PORTABLE_FLAG_FILE_NAME: &str = "ludusavi.portable";
-const MIGRATION_FLAG_FILE_NAME: &str = ".flag_migrated_legacy_config";
 
 pub type AnyError = Box<dyn std::error::Error>;
 
@@ -283,40 +282,6 @@ pub fn app_dir() -> std::path::PathBuf {
     let mut path = dirs::config_dir().unwrap();
     path.push(APP_DIR_NAME);
     path
-}
-
-/// Migrate `~/.config/ludusavi` to the newer OS-dependent location.
-///
-/// We use a flag file to prevent a spurious migration when a Linux user
-/// first launches Ludusavi with XDG_CONFIG_HOME set to default, so the
-/// `standard_app_dir` and `legacy_app_dir` happen to be the same,
-/// then later launches Ludusavi with a custom XDG_CONFIG_HOME, so the
-/// `standard_app_dir` no longer exists, but the `legacy_app_dir` does.
-pub fn migrate_legacy_config() {
-    let standard_app_dir = app_dir();
-    let mut standard_migration_flag_file = standard_app_dir.clone();
-    standard_migration_flag_file.push(MIGRATION_FLAG_FILE_NAME);
-    let mut standard_portable_flag_file = standard_app_dir.clone();
-    standard_portable_flag_file.push(PORTABLE_FLAG_FILE_NAME);
-
-    let mut legacy_app_dir = dirs::home_dir().unwrap();
-    legacy_app_dir.push(".config");
-    legacy_app_dir.push(APP_DIR_NAME);
-    let mut legacy_migration_flag_file = legacy_app_dir.clone();
-    legacy_migration_flag_file.push(MIGRATION_FLAG_FILE_NAME);
-    let mut legacy_portable_flag_file = legacy_app_dir.clone();
-    legacy_portable_flag_file.push(PORTABLE_FLAG_FILE_NAME);
-
-    if standard_app_dir.exists() && !standard_migration_flag_file.exists() && !standard_portable_flag_file.exists() {
-        let _ = std::fs::File::create(&standard_migration_flag_file);
-    } else if !standard_app_dir.exists()
-        && legacy_app_dir.exists()
-        && !legacy_migration_flag_file.exists()
-        && !legacy_portable_flag_file.exists()
-    {
-        let _ = std::fs::rename(&legacy_app_dir, &standard_app_dir);
-        let _ = std::fs::File::create(&standard_migration_flag_file);
-    }
 }
 
 #[derive(Clone, Debug, Default)]